+++ cx-portal/src/types/Constants.ts
@@ -108,7 +108,7 @@ export enum ROLES {
   SETUP_IDP = 'setup_idp',
   SETUP_CLIENT = 'setup_client',
   APPSTORE_VIEW = 'view_apps',
-  APPSTORE_VIEW_SERVICES = 'view_services',
+  APPSTORE_VIEW_SERVICES = 'view_service_marketplace',
   SUBSCRIBE_SERVICE_MARKETPLACE = 'subscribe_service_marketplace',
   APPSTORE_VIEW_DATASPACES = 'view_dataspaces',
   APPSTORE_ADD = 'add_app',